+++ desktop/src/global.d.ts
@@ -1,0 +1,36 @@
+interface Window {
+  electron: ElectronAPI;
+  api: {
+    versions: NodeJS.Process.Versions;
+    checkApiStatus: (url: string) => Promise<boolean>;
+    checkIpRange: () => Promise<boolean>;
+    getCurrentApiUrl: () => Promise<string>;
+    selectDirectory: () => Promise<string[]>;
+
+    sso: {
+      startLoginProcess: () => Promise<void>;
+      logout: (idToken: string | undefined) => Promise<void>;
+      getUser: (accessToken: string | undefined) => SSOUser<IdirIdentityProvider> | undefined;
+      fetchProtectedRoute: (
+        url: string,
+        accessToken: string | undefined,
+        options: RequestInit = {},
+      ) => Promise<[Error, null] | [null, ExtendedResponse]>;
+    };
+
+    utils: {
+      safePromise: (promise: Promise) => Promise<[Error, null] | [null, ExtendedResponse]>;
+    };
+
+    workers: {
+      copyFolderAndMetadata: ({
+        filePath,
+        transfer,
+      }: {
+        filePath: string;
+        transfer: string;
+      }) => Promise<void>;
+      getFolderMetadata: ({ filePath }: { filePath: string }) => Promise<void>;
+    };
+  };
+}
